fix: build, cleanup
