--- babel.config.json
@@ -1,19 +1,0 @@
-{
-  "presets": ["@babel/preset-env", "@babel/preset-typescript"],
-  "plugins": [
-    "@babel/plugin-transform-runtime",
-    "@babel/plugin-proposal-class-properties",
-    [
-      "module-resolver",
-      {
-        "root": ["."],
-        "alias": {
-          "utils": "./src/utils",
-          "config": "./src/config",
-          "endpoints": "./src/endpoints",
-          "types": "./src/types"
-        }
-      }
-    ]
-  ]
-}
