+++ package.json
@@ -7,27 +7,19 @@
   "author": "blockfrost.io",
   "main": "lib/index.js",
   "scripts": {
-    "build": "yarn clean && yarn tsc && babel src --out-dir lib --extensions '.ts,.tsx'",
-    "tsc": "tsc",
+    "build": "tsc",
     "clean": "rimraf lib",
-    "dev": "ts-node-dev -r tsconfig-paths/register --transpile-only src/index.ts",
     "lint": "eslint ./src/**/*.ts",
-    "generate-types": "openapi-typescript ./node_modules/blockfrost-openapi/swagger.yaml --output ./src/types/OpenApi.ts"
+    "generate-types": "openapi-typescript ./node_modules/blockfrost-openapi/swagger.yaml --output ./src/types/OpenApi.ts",
+    "run-example": "ts-node-dev --transpile-only example/index.ts"
   },
   "dependencies": {
     "axios": "^0.21.1",
+    "blockfrost-openapi": "^0.0.44",
     "rimraf": "^3.0.2",
-    "url-join": "^4.0.1",
-    "blockfrost-openapi": "^0.0.44"
+    "url-join": "^4.0.1"
   },
   "devDependencies": {
-    "openapi-typescript": "^3.0.1",
-    "@babel/cli": "^7.12.16",
-    "@babel/core": "^7.12.16",
-    "@babel/plugin-proposal-class-properties": "^7.12.13",
-    "@babel/plugin-transform-runtime": "^7.12.15",
-    "@babel/preset-env": "^7.12.16",
-    "@babel/preset-typescript": "^7.12.16",
     "@types/url-join": "^4.0.0",
     "@typescript-eslint/eslint-plugin": "^4.8.1",
     "@typescript-eslint/parser": "4.8.1",
@@ -35,10 +27,10 @@
     "eslint": "7.14.0",
     "eslint-config-airbnb-base": "14.2.1",
     "eslint-config-prettier": "6.15.0",
-    "eslint-import-resolver-babel-module": "^5.2.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-prettier": "3.1.4",
     "jest": "^26.6.3",
+    "openapi-typescript": "^3.0.1",
     "prettier": "2.2.0",
     "ts-jest": "^26.4.4",
     "ts-node": "9.0.0",
