+++ src/endpoints/blocks/index.ts
@@ -1,8 +1,8 @@
 import axios from 'axios';
-import { getHeaders, handleError } from 'utils';
-import { HashOrNumber } from 'types';
-import { components } from 'types/OpenApi';
-import { BlockFrostAPI } from '../..';
+import { getHeaders, handleError } from '../../utils';
+import { HashOrNumber } from '../../types';
+import { components } from '../../types/OpenApi';
+import { BlockFrostAPI } from '../../index';
 
 export async function blocks(
   this: BlockFrostAPI,
