+++ src/endpoints/health/index.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
-import { getHeaders, handleError } from 'utils';
-import { BlockFrostAPI } from '../..';
+import { getHeaders, handleError } from '../../utils';
+import { BlockFrostAPI } from '../../index';
 
 export function health(this: BlockFrostAPI): Promise<{ isHealthy: boolean }> {
   return new Promise((resolve, reject) => {
